Added an image class for storing the ppm data
The initial goal was to store it in a way that was printable by dumping the whole thing as bytes. This didn't work out, and the lack of const generics meant that I just used a Vec<Pixel>, where pixel contains three u8 values. util.rs was originally meant for Pixel, but I just put that in image.rs. I'll probably put non-member functions in util.rs from here on out.